Added LoadingMessage component and adjusted CSS.Added A-Z sorting for ToDo items and a button for it, although it currently does not have an API call for it.
