+++ src/components/LoadingMessage/Loading.module.css
@@ -1,0 +1,5 @@
+.LoadingMessage {
+    margin: 5px 10px 0px 10px;
+    font-family: 'Pacifico', cursive;
+    color: rgb(38, 49, 52);
+}
