+++ src/components/LoadingMessage/LoadingMessage.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import style from './Loading.module.css';
+
+const LoadingMessage = () => {
+    return(
+        <h2 className={style.LoadingMessage}>Loading....</h2>
+    )
+}
+
+export default LoadingMessage;
