+++ src/components/ToDoList.module.css
@@ -17,10 +17,11 @@
 .Header {
     color: rgb(38, 49, 52);
     background-color: rgb(254,215,102);
-    padding: 5px;
+    padding: 8px;
     margin: 0px;
     height: 3em;
     justify-content: right;
+    position: sticky;
 }
 
 .Container {
@@ -31,6 +32,14 @@
 .ListItem {
     margin: 5px;
     padding: 5px;
+}
+
+.NavItem {
+    margin: 5px 10px 0px 10px;
+    font-family: 'Pacifico', cursive;
+    color: rgb(38, 49, 52);
+    text-decoration: none;
+    font-size:x-large;
 }
 
 .AddToDoButton {
@@ -50,10 +59,19 @@
 
 .RemoveButton {
     border: 'none';
-    width: 20px;
-    height: 20px;
-    margin: 0px 5px 0px 5px;
-    align-self: flex-end;
+    width: 18px;
+    height: 18px;
+    margin: 0px 0px 0px 20px;
+}
+
+.SortToggle {
+    border: 'none';
+    margin: 8px 25% 0px 25%;
+    background-color: rgb(255,112,112);
+    color: rgb(54,73,78);
+    border-radius: 8px;
+    font-size: medium;
+    width: 6em;
 }
 
 body {
@@ -64,6 +82,10 @@ h1, h2, h3, h4 {
     font-family: 'Pacifico', cursive;
     color: rgb(38, 49, 52);
     margin: 5px;
+}
+
+link {
+    text-decoration: none;
 }
 
 ul {
